move server typings to ~/types/server
