+++ server/middlewares/validateRequests.ts
@@ -1,6 +1,7 @@
-import { INextApiMiddleware } from "~/server/types/nextApi.type";
 import { verifyJWT, verifySessionCookie } from "~/server/utils/authUtils";
 import { extractFirstQueryValue, reportBadRequest } from "~/server/utils/nextHandlerUtils";
+
+import { INextApiMiddleware } from "~/types/server/nextApi.type";
 
 /**
  * Middleware that checks whether the current user's uid and the targeted document's uid match.
