+++ server/utils/siteUtils.ts
@@ -1,6 +1,7 @@
 // TODO: Fix sites and pages models
 import { firestoreAdmin } from "~/server/firebase/firebaseAdmin";
-import { CreateSiteRequest, UpdateSiteRequest } from "~/server/types";
+
+import { CreateSiteRequest, UpdateSiteRequest } from "~/types/server";
 
 import { deleteQuery } from "./firestoreUtils";
