+++ types/server/comment.type.ts
