+++ types/server/nextApi.type.ts
