+++ types/server/page.type.ts
